[FIX] web: Readonly modifiers not properly taken into account in statusbar widget
Before this commit:
Using the readonly modifier on a statusbar widget would make it readonly
whether or not the modifier was evaluated to true.

After this commit:
The readonly attribute on a statusbar widget now work as espected and
the statusbar is readonly only when the modifier is evaluated to true.

closes odoo/odoo#123668

Opw: 3222998
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>