[FIX] purchase: Prepare move line from POL using commercial partner
When using the vendor bill autocomplete feature, the vendor bill move
lines are prepared using the values fetched from the purchase order.
However, it uses the vendor bill's partner_id instead of
commercial_partner_id and causes a mismatch in the partner_id column of
line_ids.

It is overwritten with the proper commercial_partner_id when the vendor
bill gets posted, but duplicate tax lines are not consolidated.

Now it uses commercial_partner_id from the beginning, saving the
headache of duplicate taxes on accounting reports.

closes odoo/odoo#141915

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>