[FIX] web: pdfjs: hide from all platform download & print button in attachment viewer
the download and print icon wasn't present in the document viewer.
We had to do this to avoid crash in mobile apps (Android and iOS). 

But even if there are multiple other ways to download file from desktop or laptop. some of our users are don't  see the download and print button for confidential docs. 
So in this commit, we hide it on desktop as we can now correctly handle on Android tablets and hide this button on iOS tablets.
attachements preview on screens. download and print button is hide when you open the file.